Corriger deux problèmes d'affichage liés au nombre croissant de projets
- Plan de charge, vue Année : les libellés de mois utilisaient le nom
  complet ("septembre 2026"), trop long pour les colonnes-semaine
  étroites de cette vue et tronqué en "SEPTEMBR…". Passé au format
  abrégé ("sept. 2026"), qui tient toujours dans la largeur disponible.

- Affectation des permanents : la table affichait une colonne pour
  CHAQUE projet du studio, même ceux où la ressource sélectionnée n'a
  ni réservation ni surcharge — avec plusieurs dizaines de projets,
  la plupart des colonnes étaient vides et les en-têtes (nom + budget)
  devenaient illisibles. La table ne montre plus que les projets où
  cette ressource a effectivement une réservation ou une surcharge
  manuelle, avec une largeur minimale par colonne en filet de
  sécurité.

+++ src/features/planning/timeUnits.ts
@@ -46,7 +46,9 @@ export function buildUnits(anchor: Date, zoom: ZoomLevel): TimeUnit[] {
         label: String(weekStart.getDate()),
         startIso: toISODate(weekStart),
         endIso: toISODate(weekEnd),
-        monthGroupLabel: weekStart.toLocaleDateString('fr-FR', { month: 'long', year: 'numeric' }),
+        // Short month name — the "année" zoom's per-week columns are narrow
+        // enough that "septembre 2026" gets clipped, unlike "mois"/"semaine".
+        monthGroupLabel: weekStart.toLocaleDateString('fr-FR', { month: 'short', year: 'numeric' }),
         isToday: today >= weekStart && today <= weekEnd,
         isWeekend: false,
         widthPx: WIDTH.annee,
